Split AgentHarness controllers by runtime
