+++ go/api/v1alpha2/agentharness_types.go
@@ -301,6 +301,7 @@ const (
 	AgentHarnessConditionTypeAccepted           = "Accepted"
 	AgentHarnessConditionTypeActorTemplateReady = "ActorTemplateReady"
 	AgentHarnessConditionTypeActorReady         = "ActorReady"
+	AgentHarnessConditionTypeBootstrapReady     = "BootstrapReady"
 )
 
 // +kubebuilder:object:root=true
